+++ source/12_template.sh
@@ -1,4 +1,4 @@
 bb-template() {
     local TEMPLATE="$1"
-    eval "echo \"$( < $TEMPLATE )\""
+    eval "echo \"$( < "$TEMPLATE" )\""
 }